src/requirements/test.txt
factory-boy==2.11.1
flake8==3.6.0
pytest-cov==4.0.0
pytest-django==3.6.0
